% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisel.R
\name{chisel}
\alias{chisel}
\title{Chisel}
\usage{
chisel(
  html = NULL,
  color = NULL,
  bg_color = NULL,
  duration = NULL,
  center_vertical = NULL,
  center_horizontal = NULL,
  ids = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{html}{Html content to display instead of error messages,
if \code{NULL} the original message is displyed.}

\item{color}{Color of error message text.}

\item{bg_color}{Background color of error message overlay.
If \code{NULL} will be transparent.}

\item{duration}{Duration of animation showing the error message, set to \code{0} to have none.}

\item{center_vertical}{Whether to center the message vertically and horizontally, a \code{boolean}.}

\item{center_horizontal}{Whether to center the message vertically and horizontally, a \code{boolean}.}

\item{ids}{Ids of elements to apply the cleave to. If \code{NULL} applies to all error messages.}

\item{session}{A valid shiny session.}
}
\value{
None
}
\description{
Customise silent error messages: messages raised by \code{\link[shiny:validate]{shiny::validate()}} and \code{\link[shiny:validate]{shiny::need()}}.
If used in combination with \code{\link[=cleave]{cleave()}} make sure you are setting \code{silent_errors} to \code{FALSE}.
}
