% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleave.R
\name{cleave_theme}
\alias{cleave_theme}
\title{Cleave Theme}
\usage{
cleave_theme(color = "red", bg_color = NULL, duration = 0.3, center = TRUE)
}
\arguments{
\item{color}{Color of error message text.}

\item{bg_color}{Background color of error message overlay.
If \code{NULL} will be transparent.}

\item{duration}{Duration of animation showing the error message, set to \code{0} to have none.}

\item{center}{Whether to center the message vertically and horizontally, a \code{boolean}.}
}
\description{
Define a theme to apply to all subsequent \code{\link{cleave}}.
}
