% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{dependencies}
\alias{dependencies}
\alias{useSever}
\alias{use_sever}
\title{Dependencies}
\usage{
useSever()

use_sever()
}
\value{
\link[shiny:reexports]{shiny::tags} containing the necessary dependencies.
}
\description{
Import dependencies, place this in your shiny UI.
}
