% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_golem_option}
\alias{get_golem_option}
\title{Get Golem Option}
\usage{
get_golem_option()
}
\value{
The golem prod option, if not set returns \code{TRUE}.
}
\description{
Get golem option to use whether or not to use \code{cleave} and \code{sever}.
}
\keyword{internal}
