% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theme}
\alias{get_theme}
\title{Get Cleave Theme}
\usage{
get_theme(option, value)
}
\arguments{
\item{option}{Name of the theme option to fetch.}

\item{value}{Value to return if \code{value} is \code{NULL}.}
}
\value{
A character string.
}
\description{
Returns the \code{option} unless the \code{value} is set.
}
\keyword{internal}
