% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture.R
\name{reconnect}
\alias{reconnect}
\alias{reconnect_button}
\alias{reconnect_link}
\alias{f7_reconnect_button}
\title{Reload}
\usage{
reconnect_button(
  text = "reconnect",
  class = c("default", "danger", "info", "success", "warning")
)

reconnect_link(
  text = "reconnect",
  class = c("default", "danger", "info", "success", "warning")
)

f7_reconnect_button(text = "reconnect", color = "#000")
}
\arguments{
\item{text}{The text to use on the button.}

\item{class}{The class to apply to the button.}

\item{color}{Color of button.}
}
\value{
A button or link in the form of \link[shiny:reexports]{shiny::tags}.
}
\description{
Create a button to reload/reconnect to shiny.
}
\section{Functions}{

\itemize{
\item{\code{reconnect_button} - Returns a button.}
\item{\code{reconnect_link} - Returns a link.}
\item{\code{f7_reconnect_button} - A reconnect button for shinyMobile.}
}
}

