% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sever.R
\name{reload_button}
\alias{reload_button}
\alias{reload_link}
\title{Reload}
\usage{
reload_button(
  text = "reload",
  class = c("default", "danger", "info", "success", "warning")
)

reload_link(
  text = "reload",
  class = c("default", "danger", "info", "success", "warning")
)
}
\arguments{
\item{text}{The text to use on the button.}

\item{class}{The class to apply to the button.}
}
\description{
Create a button to reload shiny.
}
