% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture.R
\name{rupture_default}
\alias{rupture_default}
\title{Default Rupture Screen}
\usage{
rupture_default(
  title = "Idle",
  subtitle = "Your session is disconnected",
  button = "Reconnect",
  button_class = "default"
)
}
\arguments{
\item{title}{Title and subtitle to display.}

\item{subtitle}{Title and subtitle to display.}

\item{button}{Text to display on button, passed to \code{\link[=reload_button]{reload_button()}}.}

\item{button_class}{Class of button, passed to \code{\link[=reload_button]{reload_button()}}.}
}
\value{
\code{shiny::tags}.
}
\description{
The default rupture screen for convenience.
}
