% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sever.R
\name{sever_default}
\alias{sever_default}
\title{Default Sever Screen}
\usage{
sever_default(
  title = "Whoops!",
  subtitle = "You have been disconnected",
  button = "Reload",
  button_class = "default"
)
}
\arguments{
\item{title, subtitle}{Title and subtitle to display.}

\item{button}{Text to display on button, passed to \code{\link[=reload_button]{reload_button()}}.}

\item{button_class}{Class of button, passed to \code{\link[=reload_button]{reload_button()}}.}
}
\value{
\code{shiny::tags}.
}
\description{
The default sever screen for convenience.
}
