% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{dependencies}
\alias{dependencies}
\alias{use_sever}
\title{Dependnecies}
\usage{
use_sever()
}
\value{
\code{shiny::tags} containing the necessary dependencies.
}
\description{
Import dependencies, place this in your shiny UI.
}
