% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sever.R
\name{reload_button}
\alias{reload_button}
\alias{reload_link}
\alias{f7_reload_button}
\title{Reload}
\usage{
reload_button(
  text = "reload",
  class = c("default", "danger", "info", "success", "warning")
)

reload_link(
  text = "reload",
  class = c("default", "danger", "info", "success", "warning")
)

f7_reload_button(text = "reload", color = "#000")
}
\arguments{
\item{text}{The text to use on the button.}

\item{class}{The class to apply to the button.}

\item{color}{Color of button.}
}
\value{
A button or link in the form of \code{shiny::tags}.
}
\description{
Create a button to reload/reconnect to shiny.
}
\section{Functions}{

\itemize{
\item{\code{reload_button} - Returns a button.}
\item{\code{reload_link} - Returns a link.}
\item{\code{f7_reload_button} - A reload button for shinyMobile.}
}
}

