% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sever.R
\name{sever}
\alias{sever}
\title{Sever}
\usage{
sever(
  ...,
  bg_color = "#333e48",
  color = "#fff",
  opacity = 1,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{...}{Shiny tags to use as content for the disconnected screen.}

\item{bg_color, color}{Background color, color of text.}

\item{opacity}{Opacity of background.}

\item{session}{A valid shiny session.}
}
\description{
Customise the Shiny disconnected screen.
}
\examples{
library(shiny)

ui <- fluidPage(
 use_sever(),
 h1("sever")
)

server <- function(input, output){
 sever(
   tagList(
     h1("Whoops"),
     reload_button()
   )
 )
}

if(interactive())
 shinyApp(ui, server)

}
